python_ai/requirements.txt
# ============================================================================
# DataPizza AI Framework - Requirements
# ============================================================================

# Core DataPizza AI Framework
datapizza-ai>=0.1.0
datapizza-ai-clients-google>=0.1.0

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Environment & Configuration
python-dotenv==1.0.1

# Vector Store
qdrant-client==1.12.1

# Google AI
google-generativeai==0.8.3

# Utilities
httpx==0.28.1
aiofiles==24.1.0

# Observability & Tracing (OpenTelemetry)
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0

# Logging
python-json-logger==3.2.1

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
